Move TransactionValidationSpec to proper package
