+++ node/src/test/scala/co/topl/modifier/transaction/AssetTransferSpec.scala
@@ -1,8 +1,5 @@
 package co.topl.modifier.transaction
 
-import cats.data.NonEmptyChain
-import co.topl.attestation.{Address, PublicKeyPropositionCurve25519}
-import co.topl.modifier.box.{AssetBox, Box, BoxId}
 import co.topl.utils.{CoreGenerators, ValidGenerators}
 import org.scalatest.EitherValues
 import org.scalatest.matchers.should.Matchers
