+++ node/src/test/scala/co/topl/modifier/transaction/PolyTransferSpec.scala
@@ -5,8 +5,6 @@ import org.scalatest.matchers.should.Matchers
 import org.scalatest.propspec.AnyPropSpec
 import org.scalatestplus.scalacheck.{ScalaCheckDrivenPropertyChecks, ScalaCheckPropertyChecks}
 
-import scala.util.Failure
-
 class PolyTransferSpec
     extends AnyPropSpec
     with ScalaCheckPropertyChecks
